srcs/app/requirements.txt
Django==5.1
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
whitenoise==6.7.0
python-dotenv==1.0.1
